feat: renomeia botão esqueci senha
- renomeia o botão e altera para que possa ser usado em outros lugares
